fetch: implementacion de borrado logico de suscribers

+++ src/controller/suscription/subscribers.controller.js
@@ -9,22 +9,22 @@ const getAllSusribers = async (req, res) => {
   res.status(200).json(suscribers);
 };
 
-const deleteSuscriptionById = async (req, res) => {
-  const { id } = req.params;
-  try {
-    await suscription.findByIdAndDelete(id);
-    res.status(200).json({ message: `Eliminado success` });
-  } catch (err) {
-    handlerHttpError(res, `origen: delete_suscription ${err}`);
-  }
-};
+// const deleteSuscriptionById = async (req, res) => {
+//   const { id } = req.params;
+//   try {
+//     await suscription.findByIdAndDelete(id);
+//     res.status(200).json({ message: `Eliminado success` });
+//   } catch (err) {
+//     handlerHttpError(res, `origen: delete_suscription ${err}`);
+//   }
+// };
 
 const addSuscription = async (req, res) => {
   const { email } = req.body;
 
   try {
     let existentSuscription = await suscription.findOne({ email: email });
-    if (!existentSuscription) {
+    if (!existentSuscription ) {
       let newSuscription = new suscription({
         email: email,
       });
@@ -52,12 +52,22 @@ const addSuscription = async (req, res) => {
         res.status(200).json({ message: "succesful_aggregate" });
       }
     } else {
-      handlerHttpError(res, "El correo ingresado ya esta suscrito", 400);
+      if(existentSuscription && existentSuscription.deleted === true){
+        existentSuscription.deleted = false
+        await existentSuscription.save()
+        let existentPerson = await person.findOne({ email: email });
+        existentPerson.suscriber = true
+        await existentPerson.save()
+        let newSuscrip = await suscription.findOne({ email: email });
+        let newEmail = newSuscrip.email;
+        await addSuscriptiontoList({ email: newEmail });
+        res.status(200).json({message: `Succes_User_activate`})
+      }
     }
   } catch (error) {
     handlerHttpError(
       res,
-      "No pudo agregarse la suscripcion o el email ingresado ya existe",
+      "No pudo agregarse la suscripcion ",
       400
     );
   }
@@ -65,7 +75,6 @@ const addSuscription = async (req, res) => {
 
 const unsuscribeUser = async (req, res) => {
   const { email } = req.query;
-  const listIds = 3;
 
   try {
     if (!email) {
@@ -79,7 +88,7 @@ const unsuscribeUser = async (req, res) => {
     let personUnsuscribe = await person.findOne({ email: email });
 
     if (!personUnsuscribe) {
-      handlerHttpError(
+     return handlerHttpError(
         res,
         "No se encontró una persona con ese correo electrónico",
         404
@@ -91,8 +100,15 @@ const unsuscribeUser = async (req, res) => {
 
     let deleteSuscriber = await suscription.findOne({ email: email });
 
-    if (deleteSuscriber) {
-      await deleteSuscriber.remove();
+    if (deleteSuscriber && deleteSuscriber.deleted === false) {
+      await suscription.findOneAndUpdate(
+        {email: email},
+        {
+          $set: {
+            deleted: true
+          }
+        }
+      );
 
       let apiInstance = ContactsApi;
       await apiInstance.deleteContact(email);
@@ -110,6 +126,6 @@ const unsuscribeUser = async (req, res) => {
 module.exports = {
   addSuscription,
   getAllSusribers,
-  deleteSuscriptionById,
+  // deleteSuscriptionById,
   unsuscribeUser,
 };
